+++ src/evaluator.rs
@@ -1,12 +1,12 @@
 use crate::{
     crd::{Policy, PolicySpec},
     events::{EventSender, PolicyEvent, PolicyEventData},
-    policy::{PolicyInfo, PolicyStoreRef},
+    policy::{PolicyInfo, PolicyStoreRef}, util::k8s_util::find_k8s_resource_matches,
 };
-use kube::core::{
+use kube::{core::{
     admission::{self, Operation},
     DynamicObject,
-};
+}, Client};
 use lazy_static::lazy_static;
 use prometheus::{register_counter_vec, CounterVec};
 use pyo3::prelude::*;
@@ -204,7 +204,7 @@ impl PolicyEvaluator {
     }
 }
 
-pub fn validate_policy_admission(
+pub async fn validate_policy_admission(
     request: &admission::AdmissionRequest<Policy>,
 ) -> (bool, Option<String>) {
     if let Some(policy) = request.object.as_ref() {
@@ -212,13 +212,32 @@ pub fn validate_policy_admission(
             Some(name) => name.as_str(),
             None => "-invalidname-",
         };
-        validate_policy(name, &policy.spec)
+        validate_policy(name, &policy.spec).await
     } else {
         (false, Some("No rule found".to_string()))
     }
 }
 
-pub fn validate_policy(name: &str, policy: &PolicySpec) -> (bool, Option<String>) {
+pub async fn validate_policy(name: &str, policy: &PolicySpec) -> (bool, Option<String>) {
+    let client = Client::try_default()
+        .await
+        .expect("failed to create kube client");
+    
+    // Iterate through match items and check whether specified resources exist in the cluster
+    for match_item in policy.target.matches.iter() {
+        println!("Now checking {} {}", &match_item.api_group, &match_item.kind);
+        let api_resource_exists = match find_k8s_resource_matches(&match_item.api_group, &match_item.kind, &client).await {
+            Ok(resources) => {
+                !resources.is_empty()
+            },
+            Err(_) => false
+        };
+
+        if ! api_resource_exists {
+            return (false, Some(format!("Specified target {}/{} is not available", match_item.api_group, match_item.kind)))
+        }
+    }
+    
     let python_code = policy.rule.python.clone();
     Python::with_gil(|py| {
         if let Err(err) = PyModule::from_code(py, &python_code, "rule.py", "bridgekeeper") {
