+++ src/util/mod.rs
@@ -1,5 +1,6 @@
 pub mod cert;
 pub mod error;
 pub mod k8s_client;
+pub mod k8s_util;
 pub mod webhook;
 pub mod defaults;
